+++ src/xpn_client/xpni/xpni_ft/xpni_ft_read_fail_r5o.c
@@ -31,8 +31,8 @@
 
         /* debugging */
         #if defined(XPNI_DEBUG)
-            printf("[%s:%d] xpni_ft_sread_r5o(%d,%p,%d,%d); \n",
-                   __FILE__,__LINE__,fd,buffer,(int)offset,size);
+            printf("[%s:%d] xpni_ft_sread_r5o(%d,%p,%d,%lu); \n",
+                   __FILE__,__LINE__,fd,buffer,(int)offset,(unsigned long)size);
         #endif
