+++ src/xpn_client/xpni/xpni_ft/xpni_ft_read_nofail_r5o.c
@@ -17,8 +17,8 @@
       {
         /* debugging */
         #if defined(XPNI_DEBUG)
-            printf("[%s:%d] xpni_ft_sread_r5o(%d,%p,%d,%d); \n",
-                   __FILE__,__LINE__,fd,buffer,(int)offset,size);
+            printf("[%s:%d] xpni_ft_sread_r5o(%d,%p,%lu,%lu); \n",
+                   __FILE__,__LINE__,fd,buffer,(unsigned long)offset,(unsigned long)size);
         #endif
 
 	return xpni_lowfsi_pread(xpni_fit_get_XPN_DATA_FD(fd),
