+++ src/xpn_client/xpni/xpni_ft/xpni_ft_write_nofail_r5o.c
@@ -51,8 +51,8 @@
 
         /* debugging */
         #if defined(XPNI_DEBUG)
-            printf("[%s:%d] xpni_ft_swrite_r5o(%d,%p,%d,%d); \n",
-                   __FILE__,__LINE__,fd,buffer,(int)offset,size);
+            printf("[%s:%d] xpni_ft_swrite_r5o(%d,%p,%d,%lu); \n",
+                   __FILE__,__LINE__,fd,buffer,(int)offset,(unsigned long)size);
         #endif
